+++ internal/command/put_int.go
@@ -8,7 +8,6 @@ func putIntCommand() *cobra.Command {
 	return &cobra.Command{
 		Use:   "int -f <file> -s <selector> <value>",
 		Short: "Update an int property in the given document.",
-		Args:  cobra.MinimumNArgs(1),
 		RunE: func(cmd *cobra.Command, args []string) error {
 			return runGenericPutCommand(genericPutOptions{
 				ValueType: "int",