+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- Nothing yet.
+### Added
+
+- `-v`, `--value` flag to workaround [dash issue](https://github.com/TomWright/dasel/issues/117).
 
 ## [v1.19.0] - 2021-08-14